// protocol-designer/src/constants.d.ts
import { LabwareDefinition2, DeckSlot as DeckDefSlot } from '@opentrons/shared-data';
import type { DeckSlot, WellVolumes } from './types';
import type { ModuleType, ModuleModel } from '@opentrons/shared-data';
export declare const getMaxVolumes: (def: LabwareDefinition2) => WellVolumes;
/** All wells for labware, in arbitrary order. */
export declare function getAllWellsForLabware(def: LabwareDefinition2): string[];
export declare const FIXED_TRASH_ID: 'fixedTrash';
export declare const STD_SLOT_X_DIM = 128;
export declare const STD_SLOT_Y_DIM = 86;
export declare const STD_SLOT_DIVIDER_WIDTH = 4;
export declare const SPAN7_8_10_11_SLOT: 'span7_8_10_11';
export declare const TC_SPAN_SLOTS: DeckSlot[];
export declare const PSEUDO_DECK_SLOTS: Record<DeckSlot, DeckDefSlot>;
export declare const START_TERMINAL_TITLE = "STARTING DECK STATE";
export declare const END_TERMINAL_TITLE = "FINAL DECK STATE";
export declare const INITIAL_DECK_SETUP_STEP_ID = "__INITIAL_DECK_SETUP_STEP__";
export declare const DEFAULT_CHANGE_TIP_OPTION: 'always';
export declare const DEFAULT_MM_FROM_BOTTOM_ASPIRATE = 1;
export declare const DEFAULT_MM_FROM_BOTTOM_DISPENSE = 0.5;
export declare const DEFAULT_MM_TOUCH_TIP_OFFSET_FROM_TOP = -1;
export declare const DEFAULT_MM_BLOWOUT_OFFSET_FROM_TOP = 0;
export declare const DEFAULT_DELAY_SECONDS = 1;
export declare const DEFAULT_WELL_ORDER_FIRST_OPTION: 't2b';
export declare const DEFAULT_WELL_ORDER_SECOND_OPTION: 'l2r';
export declare const MIN_ENGAGE_HEIGHT_V1 = -5;
export declare const MAX_ENGAGE_HEIGHT_V1 = 40;
export declare const MIN_ENGAGE_HEIGHT_V2 = -4;
export declare const MAX_ENGAGE_HEIGHT_V2 = 19;
export declare const MIN_TEMP_MODULE_TEMP = 4;
export declare const MAX_TEMP_MODULE_TEMP = 95;
export declare const MIN_HEATER_SHAKER_MODULE_TEMP = 37;
export declare const MAX_HEATER_SHAKER_MODULE_TEMP = 95;
export declare const MIN_HEATER_SHAKER_MODULE_RPM = 200;
export declare const MAX_HEATER_SHAKER_MODULE_RPM = 3000;
export declare const MIN_HEATER_SHAKER_DURATION_SECONDS = 0;
export declare const MAX_HEATER_SHAKER_DURATION_SECONDS = 60;
export declare const MIN_TC_BLOCK_TEMP = 4;
export declare const MAX_TC_BLOCK_TEMP = 99;
export declare const MIN_TC_LID_TEMP = 37;
export declare const MAX_TC_LID_TEMP = 110;
export declare const MIN_TC_DURATION_SECONDS = 0;
export declare const MAX_TC_DURATION_SECONDS = 60;
export declare const MIN_TC_PROFILE_VOLUME = 0;
export declare const MAX_TC_PROFILE_VOLUME = 100;
export declare const MODELS_FOR_MODULE_TYPE: Record<ModuleType, Array<{
    name: string;
    value: ModuleModel;
    disabled?: boolean;
}>>;
export declare const DEFAULT_MODEL_FOR_MODULE_TYPE: Record<ModuleType, ModuleModel>;
export declare const PAUSE_UNTIL_RESUME: 'untilResume';
export declare const PAUSE_UNTIL_TIME: 'untilTime';
export declare const PAUSE_UNTIL_TEMP: 'untilTemperature';
export declare const DND_TYPES: {
    LABWARE: string;
    STEP_ITEM: string;
};
export declare const THERMOCYCLER_STATE: 'thermocyclerState';
export declare const THERMOCYCLER_PROFILE: 'thermocyclerProfile';